Commit LUCY-79, making Lucy C++ compatible and changing MSVC to build in C++ mode by default.
git-svn-id: https://svn.apache.org/repos/asf/lucene/lucy/trunk@886285 13f79535-47bb-0310-9956-ffa450edef68

+++ core/Lucy/Store/RAMFileHandle.c
@@ -152,7 +152,7 @@ RAMFH_grow(RAMFileHandle *self, i64_t len)
         return false;
     }
     else {
-        BB_Grow(self->ram_file->contents, len);
+        BB_Grow(self->ram_file->contents, (size_t)len);
         return true;
     }
 }
